correct somme errors in Test classes after modifing the api core

+++ MoodleXmlApiCore/src/main/java/m2glre/marsupilami/moodlexmlapi/core/data/impl/Answer.java
@@ -71,7 +71,7 @@ public class Answer {
 	/**
 	 * @param fraction setting the fraction of the answer
 	 */
-	public void setFraction(int fraction) {
+	public void setFraction(float fraction) {
 		this.ffraction = fraction;
 	}
